+++ backend/canisters/cycles_dispenser/CHANGELOG.md
@@ -5,6 +5,8 @@ The format is based on [Keep a Changelog](https://keepachangelog.com/en/1.0.0/).
 
 ## [unreleased]
 
+## [[2.0.750](https://github.com/open-chat-labs/open-chat/releases/tag/v2.0.750-cycles_dispenser)] - 2023-07-20
+
 ### Changed
 
 - Call into ICP ledger via the new `icp_ledger_canister_c2c_client` ([#3966](https://github.com/open-chat-labs/open-chat/pull/3966))